# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aios Practice Platform API
  version: 1.0.0
  description: 题练通一期正式 OpenAPI 基线。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /roles:
    get:
      tags: [RBAC]
      summary: 角色列表
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [RBAC]
      summary: 创建角色
      requestBody:
        $ref: '#/components/requestBodies/RoleBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [RBAC]
      summary: 更新角色
      requestBody:
        $ref: '#/components/requestBodies/RoleBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    delete:
      tags: [RBAC]
      summary: 删除角色
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /permissions:
    get:
      tags: [RBAC]
      summary: 权限列表
      parameters:
        - in: query
          name: module
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [RBAC]
      summary: 创建权限
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /roles/{id}/permissions:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [RBAC]
      summary: 分配角色权限
      requestBody:
        $ref: '#/components/requestBodies/RolePermissionsBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /practice/sessions:
    get:
      tags: [Practice]
      summary: 练题记录列表
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, finished]
        - in: query
          name: course_id
          schema:
            type: integer
            format: int64
        - in: query
          name: flow_mode
          schema:
            type: string
            enum: [fixed_count, continuous]
        - in: query
          name: practice_mode
          schema:
            type: string
            enum: [random, sequential]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionPageOk' }
    post:
      tags: [Practice]
      summary: 创建练题会话
      requestBody:
        $ref: '#/components/requestBodies/PracticeSessionBody'
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionOk' }
  /practice/sessions/from-questions:
    post:
      tags: [Practice]
      summary: 从题目列表创建练题会话
      requestBody:
        $ref: '#/components/requestBodies/PracticeSessionFromQuestionsBody'
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionOk' }
  /practice/sessions/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Practice]
      summary: 练题会话详情
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionOk' }
  /practice/sessions/{id}/results:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Practice]
      summary: 练题会话结果详情
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionResultsOk' }
  /practice/sessions/{id}/next-question:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Practice]
      summary: 获取下一题
      responses:
        '200': { $ref: '#/components/responses/PracticeNextQuestionOk' }
  /practice/sessions/{id}/answer:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Practice]
      summary: 提交练题答案
      requestBody:
        $ref: '#/components/requestBodies/PracticeAnswerBody'
      responses:
        '200': { $ref: '#/components/responses/PracticeAnswerOk' }
  /practice/sessions/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Practice]
      summary: 结束练题会话
      responses:
        '200': { $ref: '#/components/responses/PracticeSessionSummaryOk' }
  /practice/questions/{id}/mark-mastered:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Practice]
      summary: 标记熟题
      requestBody:
        $ref: '#/components/requestBodies/QuestionStateBody'
      responses:
        '200': { $ref: '#/components/responses/UserQuestionStateOk' }
  /practice/questions/{id}/mark-confused:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Practice]
      summary: 标记疑惑题
      requestBody:
        $ref: '#/components/requestBodies/QuestionStateBody'
      responses:
        '200': { $ref: '#/components/responses/UserQuestionStateOk' }
  /user-question-states:
    get:
      tags: [Practice]
      summary: 用户题目状态列表
      parameters:
        - in: query
          name: state_type
          schema:
            type: string
            enum: [wrong, mastered, confused]
        - in: query
          name: bank_id
          schema:
            type: integer
            format: int64
        - in: query
          name: course_id
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        '200': { $ref: '#/components/responses/UserQuestionStatePageOk' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
...
    PracticeSessionOk:
      description: 练题会话详情
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionResponse'
    PracticeSessionPageOk:
      description: 练题会话分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionPageResponse'
    PracticeSessionResultsOk:
      description: 练题会话结果详情
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionResultsResponse'
    PracticeNextQuestionOk:
      description: 下一题结果
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeNextQuestionResponse'
    PracticeAnswerOk:
      description: 练题答题结果
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeAnswerResponse'
    PracticeSessionSummaryOk:
      description: 练题会话结束摘要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionSummaryResponse'
    UserQuestionStateOk:
      description: 用户题目状态
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserQuestionStateResponse'
    UserQuestionStatePageOk:
      description: 用户题目状态分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserQuestionStatePageResponse'
...
  requestBodies:
    RoleBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RoleInput'
    RolePermissionsBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RolePermissionsInput'
...
    PracticeSessionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionInput'
    PracticeSessionFromQuestionsBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeSessionFromQuestionsInput'
    PracticeAnswerBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PracticeAnswerInput'
    QuestionStateBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/QuestionStateInput'
...
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        request_id:
          type: string
...
    RoleInput:
      type: object
      required: [code, name, role_type, data_scope_type]
      properties:
        code:
          type: string
        name:
          type: string
        role_type:
          type: string
        data_scope_type:
          type: string
        remark:
          type: string
          nullable: true
    RolePermissionsInput:
      type: object
      required: [permission_ids]
      properties:
        permission_ids:
          type: array
          items:
            type: integer
            format: int64
...
    QuestionContentInput:
      type: object
      required: [stem]
      properties:
        stem:
          $ref: '#/components/schemas/QuestionContentBlock'
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
        option_order_randomizable:
          type: boolean
        ext:
          type: object
          additionalProperties: true
    QuestionContentBlock:
      type: object
      properties:
        content_type:
          type: string
        text:
          type: string
          nullable: true
        assets:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAsset'
...
    QuestionAnalysisInput:
      type: object
      additionalProperties: true
...
    PracticeSessionInput:
      type: object
      required: [source_mode, bank_ids]
      properties:
        practice_mode:
          type: string
          enum: [random, sequential]
        source_mode:
          type: string
          enum: [single_bank, multi_bank, course, question_list]
          description: source_mode=course 时需要传 course_id；bank_ids 可为空数组
        flow_mode:
          type: string
          enum: [fixed_count, continuous]
        course_id:
          type: integer
          format: int64
          nullable: true
          description: source_mode=course 时必填
        bank_ids:
          type: array
          items:
            type: integer
            format: int64
          description: source_mode=course 时可为空数组
        exclude_mastered:
          type: boolean
        question_count:
          type: integer
        random_seed:
          type: integer
          format: int64
    PracticeSessionFromQuestionsInput:
      type: object
      required: [question_ids]
      properties:
        question_ids:
          type: array
          items:
            type: integer
            format: int64
        practice_mode:
          type: string
          enum: [random, sequential]
        flow_mode:
          type: string
          enum: [fixed_count, continuous]
        question_count:
          type: integer
        exclude_mastered:
          type: boolean
        random_seed:
          type: integer
          format: int64
    PracticeSessionQuestion:
      type: object
      required: [session_question_id, session_id, question_id, question_version_id, display_order, question_type, content, round_no]
      properties:
        session_question_id:
          type: integer
          format: int64
        session_id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        question_version_id:
          type: integer
          format: int64
        display_order:
          type: integer
        question_type:
          type: string
        content:
          $ref: '#/components/schemas/QuestionContentInput'
        analysis:
          $ref: '#/components/schemas/QuestionAnalysisInput'
        round_no:
          type: integer
        answered:
          type: boolean
        is_correct:
          type: boolean
          nullable: true
    PracticeSessionDetail:
      type: object
      required: [id, tenant_id, user_id, practice_mode, source_mode, flow_mode, bank_scope, bank_ids, exclude_mastered, random_seed, round_no, status, questions]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        practice_mode:
          type: string
        source_mode:
          type: string
        course_id:
          type: integer
          format: int64
          nullable: true
        flow_mode:
          type: string
        bank_scope:
          type: object
          additionalProperties: true
        bank_ids:
          type: array
          items:
            type: integer
            format: int64
        exclude_mastered:
          type: boolean
        question_count:
          type: integer
        random_seed:
          type: integer
          format: int64
        round_no:
          type: integer
        status:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PracticeSessionQuestion'
    PracticeSessionListItem:
      type: object
      required: [id, status, practice_mode, source_mode, flow_mode, bank_ids, total_count, answered_count, correct_count, wrong_count, accuracy]
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        practice_mode:
          type: string
        source_mode:
          type: string
        course_id:
          type: integer
          format: int64
          nullable: true
        flow_mode:
          type: string
        bank_ids:
          type: array
          items:
            type: integer
            format: int64
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        total_count:
          type: integer
        answered_count:
          type: integer
        correct_count:
          type: integer
        wrong_count:
          type: integer
        accuracy:
          type: number
          format: double
...
    PracticeSessionResults:
      type: object
      required: [session, questions]
      properties:
        session:
          $ref: '#/components/schemas/PracticeSessionListItem'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PracticeSessionResultQuestion'
    PracticeAnswerInput:
      type: object
      required: [session_question_id, answer]
      properties:
        session_question_id:
          type: integer
          format: int64
        answer:
          type: object
          additionalProperties: true
    QuestionStateInput:
      type: object
      required: [value]
      properties:
        value:
          type: boolean
    UserQuestionState:
      type: object
      required: [id, tenant_id, user_id, question_id, question_version_id, practice_correct_count, practice_wrong_count, exam_wrong_count, is_mastered, is_confused]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        question_version_id:
          type: integer
          format: int64
        question_type:
          type: string
        content:
          type: object
          additionalProperties: true
        practice_correct_count:
          type: integer
        practice_wrong_count:
          type: integer
        exam_wrong_count:
          type: integer
        is_mastered:
          type: boolean
        is_confused:
          type: boolean
        mastered_at:
          type: string
          format: date-time
          nullable: true
        confused_at:
          type: string
          format: date-time
          nullable: true
        last_wrong_at:
          type: string
          format: date-time
          nullable: true
        last_answer:
          type: object
          additionalProperties: true
        last_result:
          type: string
        updated_at:
          type: string
          format: date-time
    PracticeAnswerResult:
      type: object
      required: [is_correct, correct_answer, state]
      properties:
        is_correct:
          type: boolean
        correct_answer:
          type: object
          additionalProperties: true
        analysis:
          $ref: '#/components/schemas/QuestionAnalysisInput'
        state:
          $ref: '#/components/schemas/UserQuestionState'
    NextPracticeQuestionResult:
      type: object
      required: [question, round_no]
      properties:
        question:
          $ref: '#/components/schemas/PracticeSessionQuestion'
        round_no:
          type: integer
    PracticeSessionSummary:
      type: object
      required: [id, status, answered_count, correct_count, wrong_count]
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        answered_count:
          type: integer
        correct_count:
          type: integer
        wrong_count:
          type: integer
...
    PracticeSessionResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/PracticeSessionDetail'
        request_id:
          type: string
    PracticeSessionPageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/PracticeSessionListItem'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
    PracticeSessionResultsResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/PracticeSessionResults'
        request_id:
          type: string
    PracticeNextQuestionResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/NextPracticeQuestionResult'
        request_id:
          type: string
    PracticeAnswerResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/PracticeAnswerResult'
        request_id:
          type: string
    PracticeSessionSummaryResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/PracticeSessionSummary'
        request_id:
          type: string
    UserQuestionStateResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserQuestionState'
        request_id:
          type: string
    UserQuestionStatePageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/UserQuestionState'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
...